requirement.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose[cryptography]
 passlib[bcrypt]
